requirements.txt
streamlit>=1.28.0
speechrecognition>=3.10.0
gTTS>=2.4.0
requests>=2.31.0
python-dotenv>=1.0.0
pyaudio>=0.2.11
openai>=1.0.0